Implement a serial console into the flashtool.
This also sets things up such that the flashtool is now capable to *configure* a device using firmware-level commands (thus a 'public stable API' between the flashtool & device).

The idea here is that with some firmware extensions (disable-captive-portal) and further work, it should be possible to completely provision a device entirely from the flashtool without using the captive portal.